Setup now installs two envs
With this change I would like to make the installer set up two enviroments, one for the stable release and one for the dev.

At the same time it is the idea that we change the `PYFUNCEBLE_CONFIG_DIR` to match the `$pyfunceblePackageName` to avoid conflicts between `pyfuncleble.yml` versions.

Todo: test the script AND have it copying the `.pyfunceble-env` from $HOME dir to miniconda/envs/